auth token realization

+++ core/api.py
@@ -1,3 +1,4 @@
+from urllib.error import HTTPError
 from rest_framework import status
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
@@ -12,39 +13,38 @@ def get_post_tickets(request) -> dict:
         tickets = Ticket.objects.all()
         data = TicketLightSerializer(tickets, many=True).data
         return Response(data=data)
-    serializer = TicketSerializer(data=request.data)
-    serializer.is_valid()
-    instance = serializer.create(serializer.validated_data)
-    results = TicketSerializer(instance).data
-    return Response(data=results, status=status.HTTP_201_CREATED)
+
+    try:
+        if request.user.auth_token:
+            serializer = TicketSerializer(data=request.data)
+            serializer.is_valid()
+            instance = serializer.create(serializer.validated_data)
+            results = TicketSerializer(instance).data
+            return Response(data=results, status=status.HTTP_201_CREATED)
+    except HTTPError:
+        data = {"details": "User has no authentetication token or token is invalid"}
+        return Response(data=data, status=status.HTTP_401_UNAUTHORIZED)
 
 
 @api_view(["GET", "PUT", "DELETE"])
 def retrieve_update_delete_ticket(request, id_: int) -> dict:
     initial_ticket = Ticket.objects.get(id=id_)
-    if request.method == "GET":
-        data = TicketSerializer(initial_ticket).data
-        return Response(data=data)
-    elif request.method == "PUT":
-        serializer = TicketSerializer(data=request.data, partial=True)
-        serializer.is_valid()
-        instance = serializer.update(initial_ticket, serializer.validated_data)
-        results = TicketSerializer(instance).data
-        return Response(data=results)
 
-    # NOTE: I made the PUT method in the same way as we did the POST method in the lesson.
-    # Code above.
-    # Basically, in the examples I met the code below (both when using the PUT method
-    # and the POST method). This edition uses the save() method instead of update() if PUT,
-    # and create() if POST. Also serialization happens only once. And checking for validity,
-    # instead of leading to validity.
+    try:
+        if request.method == "GET":
+            data = TicketSerializer(initial_ticket).data
+            return Response(data=data)
 
-    # serializer = TicketSerializer(initial_ticket, data=request.data, partial=True)
-    # if serializer.is_valid():
-    #     serializer.save()
-    #     return Response(data=serializer.data)
-    # return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        elif request.method == "PUT" and request.user.auth_token:
+            serializer = TicketSerializer(data=request.data, partial=True)
+            serializer.is_valid()
+            instance = serializer.update(initial_ticket, serializer.validated_data)
+            results = TicketSerializer(instance).data
+            return Response(data=results)
 
-    elif request.method == "DELETE":
-        initial_ticket.delete()
-        return Response({"message": "Ticket was deleted"}, status=status.HTTP_204_NO_CONTENT)
+        elif request.method == "DELETE" and request.user.auth_token:
+            initial_ticket.delete()
+            return Response({"message": "Ticket was deleted"}, status=status.HTTP_204_NO_CONTENT)
+    except HTTPError:
+        data = {"details": "User has no authentetication token or token is invalid"}
+        return Response(data=data, status=status.HTTP_401_UNAUTHORIZED)
